+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+- Add `Agilibox::Monkey`
+
 ## 1.3.4
 
 - Support FontAwesome v4